Refactor asset loading logic into super render module
